seperating article vote into a new compnent
